made reducers -- need to test functionality
