+++ src/components/AddedFeature.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { connect } from "react-redux";
 
 const AddedFeature = props => {
   return (
@@ -10,4 +11,15 @@ const AddedFeature = props => {
   );
 };
 
-export default AddedFeature;
+const mapStateToProps = state => {
+  console.log("jt: AddedFeature.js: mapPropToState: state: ", state);
+  return {
+    
+  };
+};
+
+
+export default connect(
+  mapStateToProps,
+  { }
+)(AddedFeature);
